+++ data/tests_api.py
@@ -1877,11 +1877,8 @@ class UserTestCase(APITestCase):
 class LoadQuestionnaireTestCase(APITestCase):
     def setUp(self):
         self.user_login = UserLogin(self.client)
-        add_vm_settings(self, settings_name='test-settings')
-        ShareGroup.objects.create(name='test-share-group')
-
-        self.data = {
-            "cwl_url": "https://example.org/exome-seq.cwl",
+        # Data needs to deserialize
+        self.data = {"cwl_url": "https://example.org/exome-seq.cwl",
             "workflow_version_number": 12,
             "name": "Test Questionnaire Name",
             "description" : "Test Questionnaire Description",
@@ -1920,42 +1917,20 @@ class LoadQuestionnaireTestCase(APITestCase):
         response = self.client.delete(url, format='json')
         self.assertEqual(response.status_code, status.HTTP_405_METHOD_NOT_ALLOWED)
 
-    @patch('data.api.WorkflowImporter')
-    @patch('data.api.JobQuestionnaireImporter')
-    def test_loads_questionnaire(self, mock_job_questionnaire_importer, mock_workflow_importer):
-        mock_wfi_run = Mock()
-        mock_workflow_importer.return_value.run = mock_wfi_run
-        mock_workflow_version = Mock()
-        mock_workflow_importer.return_value.workflow_version = mock_workflow_version
-        mock_jqi_run = Mock()
-        mock_job_questionnaire_importer.return_value.run = mock_jqi_run
-
+    @patch('data.api.QuestionnaireLoader')
+    def test_loads_questionnaire(self, mock_questionnaire_loader):
+        mock_run = Mock()
+        mock_questionnaire_loader.return_value.run = mock_run
         self.user_login.become_admin_user()
         url = reverse('admin_loadquestionnaire-list')
         response = self.client.post(url, self.data, format='json')
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
 
-        # Check that the workflow importer was called with the cwl_url, workflow_version_number, and methods_template_url
-        args, kwargs = mock_workflow_importer.call_args
-        self.assertEqual(args, (self.data['cwl_url'], self.data['workflow_version_number'], self.data['methods_template_url'],))
+        # Check that the Questionnaire importer was called with our POSTed data
+        args, kwargs = mock_questionnaire_loader.call_args
+        self.assertEqual(args, (self.data,))
         self.assertEqual(kwargs, {})
-        self.assertTrue(mock_wfi_run.called)
-
-        # Check that the Job Questionnaire Importer was called
-        args, kwargs = mock_job_questionnaire_importer.call_args
-        self.assertEqual(args, (
-            self.data['name'],
-            self.data['description'],
-            mock_workflow_version,
-            self.data['system_json'],
-            self.data['vm_settings_name'],
-            self.data['vm_flavor_name'],
-            self.data['share_group_name'],
-            self.data['volume_size_base'],
-            self.data['volume_size_factor'],
-        ))
-        self.assertEqual(kwargs, {})
-        self.assertTrue(mock_jqi_run.called)
+        self.assertTrue(mock_run.called)
 
     def test_unauthenticated_user_cannot_post(self):
         url = reverse('admin_loadquestionnaire-list')
@@ -1967,20 +1942,6 @@ class LoadQuestionnaireTestCase(APITestCase):
         url = reverse('admin_loadquestionnaire-list')
         response = self.client.put(url, self.data, format='json')
         self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
-
-    def test_fails_missing_vm_settings(self):
-        self.user_login.become_admin_user()
-        url = reverse('admin_loadquestionnaire-list')
-        self.data['vm_settings_name'] = 'missing-settings'
-        response = self.client.post(url, self.data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
-
-    def test_fails_missing_share_group(self):
-        self.user_login.become_admin_user()
-        url = reverse('admin_loadquestionnaire-list')
-        self.data['share_group_name'] = 'missing-share-group'
-        response = self.client.post(url, self.data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
     def test_fails_bad_system_json(self):
         # DictField on the serializer checks this
